// proyecto-flowers/src/app/models/flores.model.ts
export interface Flores {
    id: number;
    name: string;
    price: number;
    image: string;
    description: string;
    quantity?: number;
}

export interface Ramos {
    id: number;
    name: string;
    price: number;
    image: string;
    description: string;
    quantity?: number;
}

export interface Funerales {
    id: number;
    name: string;
    price: number;
    image: string;
    description: string;
    quantity?: number;
}

export interface Bodas {
    id: number;
    name: string;
    price: number;
    image: string;
    description: string;
    quantity?: number;
}